Rename utils files
