+++ kwave/kmedium.py
@@ -3,7 +3,7 @@ from typing import List
 from warnings import warn
 
 import numpy as np
-import kwave.utils.checkutils
+import kwave.utils.checks
 
 
 @dataclass
