+++ kwave/utils/__init__.py
@@ -1,15 +1,14 @@
 from kwave.utils.dotdictionary import *
-from kwave.utils.checkutils import *
+from kwave.utils.checks import *
 from kwave.utils.colormap import *
-from kwave.utils.conversionutils import *
-from kwave.utils.datautils import *
-from kwave.utils.filterutils import *
+from kwave.utils.conversion import *
+from kwave.utils.data import *
+from kwave.utils.filters import *
 from kwave.utils.interputils import *
 from kwave.utils.ioutils import *
 from kwave.utils.kutils import *
-from kwave.utils.maputils import *
-from kwave.utils.matrixutils import *
-from kwave.utils.pmlutils import *
+from kwave.utils.mapgen import *
+from kwave.utils.matrix import *
+from kwave.utils.pml import *
 from kwave.utils.tictoc import *
 from kwave.utils.math import *
-from kwave.utils.osutils import *
