--- kwave/utils/osutils.py
@@ -1,5 +1,0 @@
-import platform
-
-
-def is_unix():
-    return platform.system() in ['Linux', 'Darwin']
