+++ tests/matlab_test_data_collectors/python_testers/cart2grid_test.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 from kwave.utils.interputils import cart2grid
 
-from kwave.utils.conversionutils import scale_time
+from kwave.utils.conversion import scale_time
 
 from scipy.io import loadmat
 import numpy as np
