+++ tests/matlab_test_data_collectors/python_testers/scaleTime_test.py
@@ -1,4 +1,4 @@
-from kwave.utils.conversionutils import scale_time
+from kwave.utils.conversion import scale_time
 
 from scipy.io import loadmat
 import numpy as np
