+++ tests/test_ivp_photoacoustic_waveforms.py
@@ -14,7 +14,7 @@ from tempfile import gettempdir
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder2D import kspaceFirstOrder2DC
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
-from kwave.utils.maputils import make_ball, make_disc
+from kwave.utils.mapgen import make_ball, make_disc
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
