+++ tests/test_pr_2D_TR_line_sensor.py
@@ -13,8 +13,8 @@ from tempfile import gettempdir
 
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder2D import kspaceFirstOrder2DC
-from kwave.utils.maputils import make_disc
-from kwave.utils.filterutils import smooth
+from kwave.utils.mapgen import make_disc
+from kwave.utils.filters import smooth
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
