+++ tests/test_sd_directional_array_elements.py
@@ -13,8 +13,8 @@ from tempfile import gettempdir
 
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder2D import kspaceFirstOrder2DC
-from kwave.utils.filterutils import filter_time_series
-from kwave.utils.maputils import make_circle
+from kwave.utils.filters import filter_time_series
+from kwave.utils.mapgen import make_circle
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
