+++ tests/test_tvsp_homogeneous_medium_monopole.py
@@ -13,7 +13,7 @@ from tempfile import gettempdir
 
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder2D import kspaceFirstOrder2DC
-from kwave.utils.filterutils import filter_time_series
+from kwave.utils.filters import filter_time_series
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
