+++ tests/test_us_bmode_linear_transducer.py
@@ -12,7 +12,7 @@ import os
 from tempfile import gettempdir
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
 from kwave.utils.kutils import tone_burst
-from kwave.utils.maputils import make_ball
+from kwave.utils.mapgen import make_ball
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
