+++ tests/test_us_transducer_as_sensor.py
@@ -14,7 +14,7 @@ from tempfile import gettempdir
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
 from kwave.utils.kutils import tone_burst
-from kwave.utils.maputils import make_ball
+from kwave.utils.mapgen import make_ball
 from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
